+++ controllers/warp/configReader_test.go
@@ -1,11 +1,15 @@
 package warp
 
 import (
+	"bytes"
+	"fmt"
 	"github.com/cloudogu/k8s-service-discovery/controllers/config"
 	"github.com/cloudogu/k8s-service-discovery/controllers/warp/types"
+	"github.com/go-logr/logr/funcr"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
+	ctrl "sigs.k8s.io/controller-runtime"
 	"testing"
 )
 
@@ -15,113 +19,175 @@ func TestConfigReader_readSupport(t *testing.T) {
 		configuration: &config.Configuration{Support: []config.SupportSource{}},
 		registry:      nil,
 	}
-	t.Run("success with one disabled entry", func(t *testing.T) {
-		// given
-		expectedCategories := types.Categories{{Title: "Support", Entries: []types.Entry{
-			{Title: "aboutCloudoguToken", Target: types.TARGET_SELF, Href: "/local/href"},
-			{Title: "myCloudogu", Target: types.TARGET_EXTERNAL, Href: "https://ecosystem.cloudogu.com/"}}}}
 
-		// when
-		actual := reader.readSupport(supportSources, []string{"docsCloudoguComUrl"})
+	t.Run("should successfully read support entries without filters", func(t *testing.T) {
+		actual := reader.readSupport(supportSources, false, []string{}, []string{})
 
-		// then
-		assert.Equal(t, expectedCategories, actual, "readSupport did not return the correct Category of two entries")
-	})
-
-	t.Run("success with empty filter", func(t *testing.T) {
-		// given
 		expectedCategories := types.Categories{
 			{Title: "Support", Entries: []types.Entry{
 				{Title: "aboutCloudoguToken", Target: types.TARGET_SELF, Href: "/local/href"},
 				{Title: "myCloudogu", Target: types.TARGET_EXTERNAL, Href: "https://ecosystem.cloudogu.com/"},
-				{Title: "docsCloudoguComUrl", Target: types.TARGET_EXTERNAL, Href: "https://docs.cloudogu.com/"}}}}
-
-		// when
-		actual := reader.readSupport(supportSources, []string{})
-
-		// then
+				{Title: "docsCloudoguComUrl", Target: types.TARGET_EXTERNAL, Href: "https://docs.cloudogu.com/"},
+			}}}
 		assert.Equal(t, expectedCategories, actual)
 	})
 
-	t.Run("success with complete filter", func(t *testing.T) {
-		// given
+	t.Run("should block all entries", func(t *testing.T) {
+		actual := reader.readSupport(supportSources, true, []string{}, []string{})
+
 		expectedCategories := types.Categories{}
+		assert.Equal(t, expectedCategories, actual)
+	})
 
-		// when
-		actual := reader.readSupport(supportSources, []string{"myCloudogu", "aboutCloudoguToken", "docsCloudoguComUrl"})
+	t.Run("should add allowed entries when blocked", func(t *testing.T) {
+		actual := reader.readSupport(supportSources, true, []string{}, []string{"myCloudogu"})
 
-		// then
-		assert.Equal(t, 0, expectedCategories.Len())
-		assert.Equal(t, expectedCategories, actual, "readSupport did not return the correct Category of three entries")
+		expectedCategories := types.Categories{
+			{Title: "Support", Entries: []types.Entry{
+				{Title: "myCloudogu", Target: types.TARGET_EXTERNAL, Href: "https://ecosystem.cloudogu.com/"},
+			}}}
+		assert.Equal(t, expectedCategories, actual)
+	})
+
+	t.Run("should remove disabled entries when not blocked", func(t *testing.T) {
+		actual := reader.readSupport(supportSources, false, []string{"aboutCloudoguToken", "docsCloudoguComUrl"}, []string{})
+
+		expectedCategories := types.Categories{
+			{Title: "Support", Entries: []types.Entry{
+				{Title: "myCloudogu", Target: types.TARGET_EXTERNAL, Href: "https://ecosystem.cloudogu.com/"},
+			}}}
+		assert.Equal(t, expectedCategories, actual)
+	})
+
+	t.Run("should remove disabled entries when not blocked", func(t *testing.T) {
+		actual := reader.readSupport(supportSources, false, []string{"aboutCloudoguToken", "docsCloudoguComUrl"}, []string{})
+
+		expectedCategories := types.Categories{
+			{Title: "Support", Entries: []types.Entry{
+				{Title: "myCloudogu", Target: types.TARGET_EXTERNAL, Href: "https://ecosystem.cloudogu.com/"},
+			}}}
+		assert.Equal(t, expectedCategories, actual)
 	})
 }
 
-func TestConfigReader_getDisabledSupportIdentifiers(t *testing.T) {
-	t.Run("success", func(t *testing.T) {
-		// given
+func TestConfigReader_readStrings(t *testing.T) {
+	t.Run("should successfully read strings", func(t *testing.T) {
 		mockRegistry := newMockWatchConfigurationContext(t)
 		mockRegistry.EXPECT().Get("/config/_global/disabled_warpmenu_support_entries").Return("[\"lorem\", \"ipsum\"]", nil)
+
 		reader := &ConfigReader{
-			configuration: &config.Configuration{Support: []config.SupportSource{}},
-			registry:      mockRegistry,
+			registry: mockRegistry,
 		}
 
-		// when
-		identifiers, err := reader.getDisabledSupportIdentifiers()
-
-		// then
-		assert.Empty(t, err)
+		identifiers, err := reader.readStrings("/config/_global/disabled_warpmenu_support_entries")
+		require.NoError(t, err)
 		assert.Equal(t, []string{"lorem", "ipsum"}, identifiers)
-		mock.AssertExpectationsForObjects(t, mockRegistry)
 	})
 
-	t.Run("failed to get disabled support entries", func(t *testing.T) {
-		// given
+	t.Run("should fail reading from registry", func(t *testing.T) {
 		mockRegistry := newMockWatchConfigurationContext(t)
 		mockRegistry.EXPECT().Get("/config/_global/disabled_warpmenu_support_entries").Return("", assert.AnError)
+
 		reader := &ConfigReader{
-			configuration: &config.Configuration{Support: []config.SupportSource{}},
-			registry:      mockRegistry,
+			registry: mockRegistry,
 		}
 
-		// when
-		_, err := reader.getDisabledSupportIdentifiers()
-
-		// then
+		identifiers, err := reader.readStrings("/config/_global/disabled_warpmenu_support_entries")
 		require.Error(t, err)
-		assert.Contains(t, err.Error(), "failed to Read configuration entry /config/_global/disabled_warpmenu_support_entries from etcd")
-		mock.AssertExpectationsForObjects(t, mockRegistry)
+		assert.ErrorIs(t, err, assert.AnError)
+		assert.ErrorContains(t, err, "failed to read configuration entry /config/_global/disabled_warpmenu_support_entries from etcd")
+		assert.Equal(t, []string{}, identifiers)
 	})
 
-	t.Run("failed to unmarshal disabled support entries", func(t *testing.T) {
-		// given
+	t.Run("should fail unmarshalling", func(t *testing.T) {
 		mockRegistry := newMockWatchConfigurationContext(t)
-		mockRegistry.EXPECT().Get("/config/_global/disabled_warpmenu_support_entries").Return("{\"lorem\": \"ipsum\"}", nil)
+		mockRegistry.EXPECT().Get("/config/_global/disabled_warpmenu_support_entries").Return("not-a-string-array 123", nil)
+
 		reader := &ConfigReader{
-			configuration: &config.Configuration{Support: []config.SupportSource{}},
-			registry:      mockRegistry,
+			registry: mockRegistry,
 		}
 
-		// when
-		_, err := reader.getDisabledSupportIdentifiers()
-
-		// then
+		identifiers, err := reader.readStrings("/config/_global/disabled_warpmenu_support_entries")
 		require.Error(t, err)
-		assert.Contains(t, err.Error(), "failed to unmarshal etcd key")
-		mock.AssertExpectationsForObjects(t, mockRegistry)
+		assert.ErrorContains(t, err, "failed to unmarshal etcd key to string slice:")
+		assert.Equal(t, []string{}, identifiers)
+	})
+}
+
+func TestConfigReader_readBool(t *testing.T) {
+	t.Run("should successfully read true bool", func(t *testing.T) {
+		mockRegistry := newMockWatchConfigurationContext(t)
+		mockRegistry.EXPECT().Get("/config/_global/myBool").Return("true", nil)
+
+		reader := &ConfigReader{
+			registry: mockRegistry,
+		}
+
+		boolValue, err := reader.readBool("/config/_global/myBool")
+		require.NoError(t, err)
+		assert.True(t, boolValue)
+	})
+
+	t.Run("should successfully read false bool", func(t *testing.T) {
+		mockRegistry := newMockWatchConfigurationContext(t)
+		mockRegistry.EXPECT().Get("/config/_global/myBool").Return("false", nil)
+
+		reader := &ConfigReader{
+			registry: mockRegistry,
+		}
+
+		boolValue, err := reader.readBool("/config/_global/myBool")
+		require.NoError(t, err)
+		assert.False(t, boolValue)
+	})
+
+	t.Run("should fail reading from registry", func(t *testing.T) {
+		mockRegistry := newMockWatchConfigurationContext(t)
+		mockRegistry.EXPECT().Get("/config/_global/myBool").Return("", assert.AnError)
+
+		reader := &ConfigReader{
+			registry: mockRegistry,
+		}
+
+		boolValue, err := reader.readBool("/config/_global/myBool")
+		require.Error(t, err)
+		assert.ErrorIs(t, err, assert.AnError)
+		assert.ErrorContains(t, err, "failed to read configuration entry /config/_global/myBool from etcd")
+		assert.False(t, boolValue)
+	})
+
+	t.Run("should fail unmarshalling", func(t *testing.T) {
+		mockRegistry := newMockWatchConfigurationContext(t)
+		mockRegistry.EXPECT().Get("/config/_global/myBool").Return("not a bool", nil)
+
+		reader := &ConfigReader{
+			registry: mockRegistry,
+		}
+
+		boolValue, err := reader.readBool("/config/_global/myBool")
+		require.Error(t, err)
+		assert.ErrorContains(t, err, "failed to unmarshal etcd key to bool:")
+		assert.False(t, boolValue)
 	})
 }
 
 func TestConfigReader_readFromConfig(t *testing.T) {
-	mockRegistry := newMockWatchConfigurationContext(t)
-	mockRegistry.EXPECT().GetChildrenPaths("/path/to/etcd/key").Return([]string{"/path/to/etcd/key"}, nil)
-	mockRegistry.EXPECT().Get("/config/_global/disabled_warpmenu_support_entries").Return("[\"lorem\", \"ipsum\"]", nil)
-	testSources := []config.Source{{Path: "/path/to/etcd/key", Type: "externals", Tag: "tag"}, {Path: "/path", Type: "disabled_support_entries"}}
+
+	testSources := []config.Source{{Path: "/path/to/etcd/key", Type: "externals", Tag: "tag"}, {Path: "/path", Type: "support_entry_config"}}
 	testSupportSoureces := []config.SupportSource{{Identifier: "supportSrc", External: true, Href: "path/to/external"}}
-	mockDoguConverter := NewMockDoguConverter(t)
 
 	t.Run("success with one external and support link", func(t *testing.T) {
 		// given
+		mockRegistry := newMockWatchConfigurationContext(t)
+		mockRegistry.EXPECT().GetChildrenPaths("/path/to/etcd/key").Return([]string{"/path/to/etcd/key"}, nil)
+		mockRegistry.EXPECT().Get("/config/_global/disabled_warpmenu_support_entries").Return("[\"lorem\", \"ipsum\"]", nil)
+		mockRegistry.EXPECT().Get(blockWarpSupportCategoryConfigurationKey).
+			Return("false", nil)
+		mockRegistry.EXPECT().Get(allowedWarpSupportEntriesConfigurationKey).
+			Return("[\"lorem\", \"ipsum\"]", nil)
+
+		mockDoguConverter := NewMockDoguConverter(t)
+
 		cloudoguEntryWithCategory := getEntryWithCategory("Cloudogu", "www.cloudogu.com", "Cloudogu", "External", types.TARGET_EXTERNAL)
 		mockExternalConverter := NewMockExternalConverter(t)
 		mockExternalConverter.EXPECT().ReadAndUnmarshalExternal(mockRegistry, mock.Anything).Return(cloudoguEntryWithCategory, nil)
@@ -144,6 +210,7 @@ func TestConfigReader_readFromConfig(t *testing.T) {
 
 	t.Run("success with one dogu and support link", func(t *testing.T) {
 		// given
+
 		mockDoguConverter := NewMockDoguConverter(t)
 		mockExternalConverter := NewMockExternalConverter(t)
 		doguSource := config.Source{
@@ -154,6 +221,11 @@ func TestConfigReader_readFromConfig(t *testing.T) {
 		mockRegistry := newMockWatchConfigurationContext(t)
 		mockRegistry.EXPECT().GetChildrenPaths(mock.Anything).Return([]string{}, nil)
 		mockRegistry.EXPECT().Get("/config/_global/disabled_warpmenu_support_entries").Return("[\"lorem\", \"ipsum\"]", nil)
+		mockRegistry.EXPECT().Get(blockWarpSupportCategoryConfigurationKey).
+			Return("false", nil)
+		mockRegistry.EXPECT().Get(allowedWarpSupportEntriesConfigurationKey).
+			Return("[\"lorem\", \"ipsum\"]", nil)
+
 		reader := &ConfigReader{
 			configuration:     &config.Configuration{Support: []config.SupportSource{}},
 			registry:          mockRegistry,
@@ -173,6 +245,16 @@ func TestConfigReader_readFromConfig(t *testing.T) {
 
 	t.Run("error during external Read should not result in an error", func(t *testing.T) {
 		// given
+		mockRegistry := newMockWatchConfigurationContext(t)
+		mockRegistry.EXPECT().GetChildrenPaths("/path/to/etcd/key").Return([]string{"/path/to/etcd/key"}, nil)
+		mockRegistry.EXPECT().Get("/config/_global/disabled_warpmenu_support_entries").Return("[\"lorem\", \"ipsum\"]", nil)
+		mockRegistry.EXPECT().Get(blockWarpSupportCategoryConfigurationKey).
+			Return("false", nil)
+		mockRegistry.EXPECT().Get(allowedWarpSupportEntriesConfigurationKey).
+			Return("[\"lorem\", \"ipsum\"]", nil)
+
+		mockDoguConverter := NewMockDoguConverter(t)
+
 		mockExternalConverter := NewMockExternalConverter(t)
 		mockExternalConverter.EXPECT().ReadAndUnmarshalExternal(mockRegistry, mock.Anything).Return(types.EntryWithCategory{}, assert.AnError)
 		reader := &ConfigReader{
@@ -197,8 +279,15 @@ func TestConfigReader_readFromConfig(t *testing.T) {
 		mockRegistry := newMockWatchConfigurationContext(t)
 		mockRegistry.EXPECT().GetChildrenPaths("/path/to/etcd/key").Return([]string{"/path/to/etcd/key"}, nil)
 		mockRegistry.EXPECT().Get("/config/_global/disabled_warpmenu_support_entries").Return("[\"lorem\", \"ipsum\"]", assert.AnError)
+		mockRegistry.EXPECT().Get(blockWarpSupportCategoryConfigurationKey).
+			Return("false", nil)
+		mockRegistry.EXPECT().Get(allowedWarpSupportEntriesConfigurationKey).
+			Return("[\"lorem\", \"ipsum\"]", nil)
+
 		mockExternalConverter := NewMockExternalConverter(t)
 		mockExternalConverter.EXPECT().ReadAndUnmarshalExternal(mockRegistry, mock.Anything).Return(types.EntryWithCategory{}, assert.AnError)
+		mockDoguConverter := NewMockDoguConverter(t)
+
 		reader := &ConfigReader{
 			configuration:     &config.Configuration{Support: []config.SupportSource{}},
 			registry:          mockRegistry,
@@ -221,8 +310,14 @@ func TestConfigReader_readFromConfig(t *testing.T) {
 		mockRegistry := newMockWatchConfigurationContext(t)
 		mockRegistry.EXPECT().GetChildrenPaths("/path/to/etcd/key").Return([]string{"/path/to/etcd/key"}, nil)
 		mockRegistry.EXPECT().Get("/config/_global/disabled_warpmenu_support_entries").Return("[]", nil)
+		mockRegistry.EXPECT().Get(blockWarpSupportCategoryConfigurationKey).
+			Return("false", nil)
+		mockRegistry.EXPECT().Get(allowedWarpSupportEntriesConfigurationKey).
+			Return("[\"lorem\", \"ipsum\"]", nil)
+
 		mockExternalConverter := NewMockExternalConverter(t)
 		mockExternalConverter.EXPECT().ReadAndUnmarshalExternal(mockRegistry, mock.Anything).Return(types.EntryWithCategory{}, assert.AnError)
+		mockDoguConverter := NewMockDoguConverter(t)
 		reader := &ConfigReader{
 			configuration:     &config.Configuration{Support: []config.SupportSource{}},
 			registry:          mockRegistry,
@@ -242,6 +337,13 @@ func TestConfigReader_readFromConfig(t *testing.T) {
 
 	t.Run("skip wrong source type", func(t *testing.T) {
 		// given
+		mockRegistry := newMockWatchConfigurationContext(t)
+		mockRegistry.EXPECT().Get("/config/_global/disabled_warpmenu_support_entries").Return("[\"lorem\", \"ipsum\"]", nil)
+		mockRegistry.EXPECT().Get(blockWarpSupportCategoryConfigurationKey).
+			Return("false", nil)
+		mockRegistry.EXPECT().Get(allowedWarpSupportEntriesConfigurationKey).
+			Return("[\"lorem\", \"ipsum\"]", nil)
+
 		testSources := []config.Source{{Path: "/path/to/etcd/key", Type: "fjkhsdfjh", Tag: "tag"}}
 		reader := &ConfigReader{
 			configuration: &config.Configuration{Support: []config.SupportSource{}},
@@ -252,6 +354,91 @@ func TestConfigReader_readFromConfig(t *testing.T) {
 
 		// then
 		require.NoError(t, err)
+	})
+
+	t.Run("should read categories from config", func(t *testing.T) {
+		mockRegistry := newMockWatchConfigurationContext(t)
+		mockRegistry.EXPECT().Get(blockWarpSupportCategoryConfigurationKey).
+			Return("false", nil)
+		mockRegistry.EXPECT().Get(disabledWarpSupportEntriesConfigurationKey).
+			Return("[\"lorem\", \"ipsum\"]", nil)
+		mockRegistry.EXPECT().Get(allowedWarpSupportEntriesConfigurationKey).
+			Return("[\"lorem\", \"ipsum\"]", nil)
+		mockRegistry.EXPECT().GetChildrenPaths("/config/externals").
+			Return([]string{"/config/externals/ext1"}, nil)
+
+		mockConverter := NewMockExternalConverter(t)
+		mockConverter.EXPECT().ReadAndUnmarshalExternal(mockRegistry, "/config/externals/ext1").Return(types.EntryWithCategory{
+			Entry: types.Entry{
+				DisplayName: "ext1",
+				Href:        "https://my.url/ext1",
+				Title:       "ext1 Description",
+				Target:      types.TARGET_EXTERNAL,
+			},
+			Category: "Documentation",
+		}, nil)
+
+		reader := &ConfigReader{
+			configuration:     &config.Configuration{Support: []config.SupportSource{}},
+			registry:          mockRegistry,
+			externalConverter: mockConverter,
+		}
+
+		testSources := []config.Source{{Path: "/config/externals", Type: "externals", Tag: "tag"}}
+		testSupportSoureces := []config.SupportSource{{Identifier: "supportSrc", External: true, Href: "https://support.source"}}
+
+		actual, err := reader.Read(&config.Configuration{Sources: testSources, Support: testSupportSoureces})
+		require.NoError(t, err)
+
+		expectedCategories := types.Categories{
+			{Title: "Documentation", Entries: []types.Entry{
+				{DisplayName: "ext1", Title: "ext1 Description", Target: types.TARGET_EXTERNAL, Href: "https://my.url/ext1"},
+			}},
+			{Title: "Support", Entries: []types.Entry{
+				{Title: "supportSrc", Target: types.TARGET_EXTERNAL, Href: "https://support.source"},
+			}},
+		}
+		assert.Equal(t, expectedCategories, actual)
+	})
+
+	t.Run("should read categories from config", func(t *testing.T) {
+		mockRegistry := newMockWatchConfigurationContext(t)
+		mockRegistry.EXPECT().Get(blockWarpSupportCategoryConfigurationKey).Return("", assert.AnError)
+		mockRegistry.EXPECT().Get(disabledWarpSupportEntriesConfigurationKey).Return("", assert.AnError)
+		mockRegistry.EXPECT().Get(allowedWarpSupportEntriesConfigurationKey).Return("", assert.AnError)
+		mockRegistry.EXPECT().GetChildrenPaths("/config/externals").Return(nil, assert.AnError)
+
+		mockConverter := NewMockExternalConverter(t)
+
+		reader := &ConfigReader{
+			configuration:     &config.Configuration{Support: []config.SupportSource{}},
+			registry:          mockRegistry,
+			externalConverter: mockConverter,
+		}
+
+		testSources := []config.Source{{Path: "/config/externals", Type: "externals", Tag: "tag"}}
+		testSupportSources := []config.SupportSource{}
+
+		// capture log
+		var buf bytes.Buffer
+		existingLog := ctrl.Log
+		ctrl.Log = funcr.New(func(prefix, args string) {
+			buf.WriteString(fmt.Sprintf("[%s] %s\n", prefix, args))
+		}, funcr.Options{})
+		defer func() {
+			ctrl.Log = existingLog
+		}()
+
+		actual, err := reader.Read(&config.Configuration{Sources: testSources, Support: testSupportSources})
+		require.NoError(t, err)
+
+		assert.Nil(t, actual)
+
+		// assert log
+		assert.Contains(t, buf.String(), "Error during Read: failed to Read root entry /config/externals from etcd")
+		assert.Contains(t, buf.String(), "Warning, could not read etcd Key: /config/_global/block_warpmenu_support_category.")
+		assert.Contains(t, buf.String(), "Warning, could not read etcd Key: /config/_global/disabled_warpmenu_support_entries.")
+		assert.Contains(t, buf.String(), "Warning, could not read etcd Key: /config/_global/allowed_warpmenu_support_entries.")
 	})
 }